projects/analysis/pyproject.toml
[tool.poetry]
name = "analysis"
version = "0.0.1"
description = "Create HTML file with analysis of DeepClean outputs"
authors = ["Alec Gunny <alec.gunny@gmail.com>"]
# readme = "README.md"

[tool.poetry.scripts]
analyze = "analysis:main"

[tool.poetry.dependencies]
python = "^3.8,<3.10"
numpy = "^1.21"
scipy = "^1.7"
bokeh = "^2.3"
"hermes.typeo" = {path = "../../gw-iaas/libs/hermes/hermes.typeo", develop = true}

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry>=1.2"]
build-backend = "poetry.masonry.api"
